+++ src/Components/CreateImageWizard/steps/Oscap/components/PolicySelector.tsx
@@ -27,7 +27,6 @@ import {
   selectDistribution,
   selectCompliancePolicyID,
   selectCompliancePolicyTitle,
-  addPackage,
   addPartition,
   changeFileSystemConfigurationType,
   clearPartitions,
@@ -90,7 +89,7 @@ const PolicySelector = () => {
   const hasWslTargetOnly = useHasSpecificTargetOnly('wsl');
   const dispatch = useAppDispatch();
   const [isOpen, setIsOpen] = useState(false);
-  const { clearCompliancePackages } = useSelectorHandlers();
+  const { clearCompliancePackages, handlePackages } = useSelectorHandlers();
 
   const {
     data: policies,
@@ -147,23 +146,6 @@ const PolicySelector = () => {
     dispatch(changeFileSystemConfigurationType('automatic'));
     handleServices(undefined);
     dispatch(clearKernelAppend());
-  };
-
-  const handlePackages = (
-    oldOscapPackages: string[],
-    newOscapPackages: string[]
-  ) => {
-    clearCompliancePackages(oldOscapPackages);
-
-    for (const pkg of newOscapPackages) {
-      dispatch(
-        addPackage({
-          name: pkg,
-          summary: 'Required by chosen compliance policy',
-          repository: 'distro',
-        })
-      );
-    }
   };
 
   const handlePartitions = (oscapPartitions: Filesystem[]) => {
@@ -223,7 +205,11 @@ const PolicySelector = () => {
           const oscapPartitions = response.filesystem || [];
           const newOscapPackages = response.packages || [];
           handlePartitions(oscapPartitions);
-          handlePackages(oldOscapPackages, newOscapPackages);
+          handlePackages(
+            oldOscapPackages,
+            newOscapPackages,
+            'Required by chosen compliance policy'
+          );
           handleServices(response.services);
           handleKernelAppend(response.kernel?.append);
           dispatch(
